+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+- 优化了些 `f-button` 组件的样式细节
+- 优化 `f-toolbar` 组件的一些内部实现和类型
+
 ## 0.8.0-alpha.6 (2022-11-06)
 
 **修复 Bug**